# specs/swagger.yaml
info:
  version: 0.0.1
  title: diagnostic-au-ghm
paths:
  '/diagnostic/{code}/acts':
    x-swagger-router-controller: Diagnostic
    get:
      tags:
        - Diagnostic
      description: Get ACTs details by Diagnostic code.
      operationId: getACTsByDiagnosticCode
      parameters:
        - name: code
          in: path
          description: The Diagnostic code of acts.
          required: true
          type: string
        - name: start
          in: query
          description: The start index of acts.
          type: integer
        - name: size
          in: query
          description: The number of item of acts.
          type: integer
        - name: filters
          in: query
          description: The filter keywords of acts.
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/MedicalActsResponse'
        '403':
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorsResponse'
      consumes:
        - application/json
      produces:
        - application/json
  /diagnostic:
    x-swagger-router-controller: Diagnostic
    post:
      tags:
        - Diagnostic
      description: Diagnostic a GHM by decision tree making algorithm.
      operationId: diagnosticGHM
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DiagnosticResponse'
        '403':
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorsResponse'
      x-code-samples:
        - lang: CURL
          source: |
            curl -X POST --header 'Content-Type: application/json' --header 'Accept: application/json' -d '"clientParameters":{"age": 28,"session": 1}' 'https://localhost:3000/diagnostic'
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: The object contains diagnostic request.
          required: true
          name: body
          in: body
          schema:
            $ref: '#/definitions/DiagnosticRequest'
  '/ghms/{code}':
    x-swagger-router-controller: GHMs
    get:
      tags:
        - GHMs
      description: Get GHM details by GHM code.
      operationId: getGHMByCode
      parameters:
        - name: code
          in: path
          description: The GHM code of a GHM.
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/GHMResponse'
        '403':
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorsResponse'
      consumes:
        - application/json
      produces:
        - application/json
swagger: '2.0'
definitions:
  DiagnosticRequest:
    properties:
      clientParameters:
        type: object
        properties:
          age:
            type: integer
          sessions:
            type: integer
          gender:
            type: string
        description: The client's information
      mainDiagnosis:
        type: string
        example: CMD-28
        description: The main category diagnosis code
      associatedDiagnoses:
        type: array
        items:
          type: string
        example:
          - D-145
          - D-114
  DiagnosticResponse:
    properties:
      associatedDiagnoses:
        type: array
        items:
          type: string
        example:
          - D-113
          - D-114
          - D-063
          - D-142
          - D-143
      possibleGHMCodes:
        type: array
        items:
          type: string
        example:
          - 28Z01
          - 28Z02
          - 28Z03
          - 28Z04
          - 90Z02Z
        description: The all possible GHMs
  MedicalActsResponse:
    properties:
      medicalActs:
        type: array
        items:
          type: object
        example:
          - code: A-112
            name: Act Code 112
            desc: Medical act 112 of D-143
            refs: ''
          - code: A-113
            name: Act Code 113
            desc: Medical act 113 of D-143
            refs: ''
          - code: A-114
            name: Act Code 114
            desc: Medical act 114 of D-143
            refs: ''
  GHMResponse:
    properties:
      name:
        type: string
        description: The GHM description
        example: Seasen d aphérèse sanguine (D-054)
      serv:
        type: string
        description: 'The severity (1,2..A,B,C) of GHM 28Z16'
        example: 1
      flag:
        type: string
        description: 'The flag set (J,T) of GHM 28Z16'
        example: J
      desc:
        type: string
        description: The description of GHM 28Z16
        example: The GHM code of CMD-28
      refs:
        type: string
        description: The references
        example: ''
  ErrorsResponse:
    required:
      - code
    properties:
      code:
        type: string
      message:
        type: string
x-components: {}